backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
celery==5.4.0
redis==5.2.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.27.2
